% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm4_methods.R
\name{case.names.glm4}
\alias{case.names.glm4}
\title{Case and Variable Names of Fitted Models}
\usage{
\method{case.names}{glm4}(object, full = FALSE, ...)
}
\arguments{
\item{object}{an \R object, typically a fitted model.}

\item{full}{logical; if \code{TRUE}, all names (including zero weights,
   \dots) are returned.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A character vector.
}
\description{
Simple utilities returning (non-missing) case names, and
  (non-eliminated) variable names.
}
\examples{
x <- 1:20
y <-  setNames(x + (x/4 - 2)^3 + rnorm(20, sd = 3),
               paste("O", x, sep = "."))
ww <- rep(1, 20); ww[13] <- 0
summary(lmxy <- lm(y ~ x + I(x^2)+I(x^3) + I((x-10)^2), weights = ww),
        correlation = TRUE)
variable.names(lmxy)
variable.names(lmxy, full = TRUE)  # includes the last
case.names(lmxy)
case.names(lmxy, full = TRUE)      # includes the 0-weight case
}
\seealso{
\code{\link[stats]{lm}}; further, \code{\link{all.names}},
  \code{\link{all.vars}} for functions with a similar name but only slightly
  related purpose.
}
\keyword{internal}
