% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm4_methods.R
\name{deviance.glm4}
\alias{deviance.glm4}
\title{Model Deviance}
\usage{
\method{deviance}{glm4}(object, ...)
}
\arguments{
\item{object}{an object for which the deviance is desired.}

\item{...}{additional optional argument.}
}
\value{
A numeric scalar giving the model deviance.
}
\description{
Returns the deviance of a fitted model object.
}
\details{
This is a generic function which can be used to extract deviances for
  fitted models.  Consult the individual modeling functions for details
  on how to use this function.
}
\examples{
fit <- glm4(mpg ~ cyl + wt, data = mtcars, family = gaussian())
deviance(fit)
}
\references{
Chambers, J. M. and Hastie, T. J. (1992)
  \emph{Statistical Models in S.}
  Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[stats]{df.residual}},
  \code{\link[stats]{extractAIC}},
  \code{\link[stats]{glm}},
  \code{\link[stats]{lm}}.
}
\keyword{internal}
