% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm4_methods.R
\name{df.residual.glm4}
\alias{df.residual.glm4}
\title{Residual Degrees-of-Freedom}
\usage{
\method{df.residual}{glm4}(object)
}
\arguments{
\item{object}{an object for which the degrees-of-freedom are desired.}
}
\value{
The value of the residual degrees-of-freedom extracted from the object
  \code{x}.
}
\description{
Returns the residual degrees-of-freedom extracted from a fitted model
  object.
}
\details{
This is a generic function which can be used to extract residual
  degrees-of-freedom for fitted models.  Consult the individual modeling
  functions for details on how to use this function.

  The default method just extracts the \code{df.residual} component.
}
\seealso{
\code{\link[stats]{deviance}}, \code{\link[stats]{glm}}, \code{\link[stats]{lm}}.
}
\keyword{internal}
