% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm4_methods.R
\name{extractAIC.glm4}
\alias{extractAIC.glm4}
\title{Extract AIC from a Fitted Model}
\usage{
\method{extractAIC}{glm4}(fit, scale = 0, k = 2, ...)
}
\arguments{
\item{fit}{fitted model, usually the result of a fitter like
    \code{\link[stats]{lm}}.}

\item{scale}{optional numeric specifying the scale parameter of the
   model, see \code{scale} in \code{\link[stats]{step}}.  Currently only used
   in the \code{"lm"} method, where \code{scale} specifies the estimate
   of the error variance, and \code{scale = 0} indicates that it is to
   be estimated by maximum likelihood.
 }

\item{k}{numeric specifying the \sQuote{weight} of the
   \emph{equivalent degrees of freedom} (\eqn{\equiv}{=:} \code{edf})
   part in the AIC formula.}

\item{...}{further arguments (currently unused in base \R).}
}
\value{
A numeric vector of length 2, with first and second elements giving

  \item{edf}{the \sQuote{\bold{e}quivalent \bold{d}egrees of \bold{f}reedom}
    for the fitted model \code{fit}.}

  \item{AIC}{the (generalized) Akaike Information Criterion for \code{fit}.}
}
\description{
Computes the (generalized) Akaike \bold{A}n \bold{I}nformation
  \bold{C}riterion for a fitted parametric model.
}
\details{
This is a generic function, with methods in base \R for classes
  \code{"aov"}, \code{"glm"} and \code{"lm"} as well as for
  \code{"negbin"} (package \href{https://CRAN.R-project.org/package=MASS}{\pkg{MASS}}) and \code{"coxph"} and
  \code{"survreg"} (package \href{https://CRAN.R-project.org/package=survival}{\pkg{survival}}).

  The criterion used is
  \deqn{AIC = - 2\log L +  k \times \mbox{edf},}{AIC = - 2*log L +  k * edf,}
  where \eqn{L} is the likelihood and \code{edf} the equivalent degrees
  of freedom (i.e., the number of free parameters for usual parametric
  models) of \code{fit}.

  For linear models with unknown scale (i.e., for \code{\link[stats]{lm}} and
  \code{\link[stats]{aov}}), \eqn{-2\log L}{-2 log L} is computed from the
  \emph{deviance} and uses a different additive constant to
  \code{\link[stats]{logLik}} and hence \code{\link[stats]{AIC}}.  If \eqn{RSS}
  denotes the (weighted) residual sum of squares then \code{extractAIC}
  uses for \eqn{- 2\log L}{-2 log L} the formulae \eqn{RSS/s - n} (corresponding
  to Mallows' \eqn{C_p}{Cp}) in the case of known scale \eqn{s} and
  \eqn{n \log (RSS/n)}{n log (RSS/n)} for unknown scale.
  \code{\link[stats]{AIC}} only handles unknown scale and uses the formula
  \eqn{n \log (RSS/n) + n + n \log 2\pi - \sum \log w}{n*log(RSS/n) + n + n*log 2pi - sum(log w)}
  where \eqn{w} are the weights.  Further \code{AIC} counts the scale
  estimation as a parameter in the \code{edf} and \code{extractAIC} does not.

  For \code{glm} fits the family's \code{aic()} function is used to
  compute the AIC: see the note under \code{logLik} about the
  assumptions this makes.

  \code{k = 2} corresponds to the traditional AIC, using \code{k =
    log(n)} provides the BIC (Bayesian \abbr{IC}) instead.

  Note that the methods for this function may differ in their
  assumptions from those of methods for \code{\link[stats]{AIC}} (usually
  \emph{via} a method for \code{\link[stats]{logLik}}).  We have already
  mentioned the case of \code{"lm"} models with estimated scale, and
  there are similar issues in the \code{"glm"} and \code{"negbin"}
  methods where the dispersion parameter may or may not be taken as
  \sQuote{free}.  This is immaterial as \code{extractAIC} is only used
  to compare models of the same class (where only differences in AIC
  values are considered).
}
\note{
This function is used in \code{\link[stats]{add1}}, \code{\link[stats]{drop1}}
  and \code{\link[stats]{step}} and the similar functions in package
  \href{https://CRAN.R-project.org/package=MASS}{\pkg{MASS}} from which it was adopted.
}
\examples{
\donttest{
utils::example(glm)
extractAIC(glm.D93)  #>>  5  15.129
}
}
\references{
Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.}
  New York: Springer (4th ed).
}
\seealso{
\code{\link[stats]{AIC}}, \code{\link[stats]{deviance}}, \code{\link[stats]{add1}},
  \code{\link[stats]{step}}
}
\author{
B. D. Ripley
}
\keyword{internal}
