% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm4.R
\name{glm4}
\alias{glm4}
\title{Fitting Generalized Linear Models Using Sparse Matrices}
\usage{
glm4(formula, family, data, weights, subset, na.action,
start = NULL, etastart, mustart, offset,
sparse = FALSE, drop.unused.levels = FALSE, doFit = TRUE,
control = list(\dots),
model = TRUE, x = FALSE, y = TRUE, contrasts = NULL, \dots)
}
\arguments{
\item{formula}{an object of class \code{"\link{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{data}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{glm} is called.}

\item{...}{potentially arguments passed on to fitter functions; not
    used currently.}
}
\value{
A list object of class `glm4`. See `stats::glm()` for more details on returned components.
}
\description{
`glm4()`, is used to fit generalized linear models,  specified by giving a symbolic description of the linear predictor and a description of the error distribution.

It is very similar to the standard `stats::glm()` function, but supports sparse matrices via the Matrix package, which can dramatically improve memory and computational efficiency on large and/or high-dimensional data.
}
\details{
This function is a wrapper for `MatrixModels::glm4()` which returns an more user-friendly object designed to resemble `stats::glm()` as closely as possible. Behind the scenes, it extracts the relevant model details from the S4 class `MatrixModels::glm4()` object and calculates new ones where necessary (e.g. AIC, deviance, residual degrees of freedom, etc) as per `stats::glm()`.
}
