% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm4_methods.R
\name{nobs.glm4}
\alias{nobs.glm4}
\title{Extract the Number of Observations from a Fit}
\usage{
\method{nobs}{glm4}(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{...}{further arguments to be passed to methods.}
}
\value{
An integer giving the number of observations used in model fitting.
}
\description{
Extract the number of \sQuote{observations} from a model fit.  This is
  principally intended to be used in computing BIC (see \code{\link[stats]{AIC}}).
}
\details{
This is a generic function, with an S4 generic in package \pkg{stats4}.
  There are methods in this package for objects of classes
  \code{"\link[stats]{lm}"}, \code{"\link[stats]{glm}"}, \code{"\link[stats]{nls}"} and
  \code{"\link[stats]{logLik}"}, as well as a default method (which throws an
  error, unless \code{use.fallback = TRUE} when it looks for
  \code{weights} and \code{residuals} components -- use with care!).

  The main usage is in determining the appropriate penalty for BIC, but
  \code{nobs} is also used by the stepwise fitting methods
  \code{\link[stats]{step}}, \code{\link[stats]{add1}} and \code{\link[stats]{drop1}} as a
  quick check that different fits have been fitted to the same set of
  data (and not, say, that further rows have been dropped because of NAs
  in the new predictors).

  For \code{lm}, \code{glm} and \code{nls} fits, observations with zero
  weight are not included.
}
\examples{
fit <- glm4(mpg ~ cyl + wt, data = mtcars, family = gaussian())
nobs(fit)
}
\seealso{
\code{\link[stats]{AIC}}.
}
\keyword{internal}
