% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm4_methods.R
\name{predict.glm4}
\alias{predict.glm4}
\title{Predict Method for GLM Fits}
\usage{
\method{predict}{glm4}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\value{
If \code{se.fit = FALSE}, a vector or matrix of predictions.
  For \code{type = "terms"} this is a matrix with a column per term, and
  may have an attribute \code{"constant"}.
  
  If \code{se.fit = TRUE}, a list with components
  \item{fit}{Predictions, as for \code{se.fit = FALSE}.}
  \item{se.fit}{Estimated standard errors.}
  \item{residual.scale}{A scalar giving the square root of the
    dispersion used in computing the standard errors.}
}
\description{
Obtains predictions and optionally estimates standard errors of those
  predictions from a fitted generalized linear model object.
}
\details{
If \code{newdata} is omitted the predictions are based on the data
  used for the fit.  In that case how cases with missing values in the
  original fit is determined by the \code{na.action} argument of that
  fit.  If \code{na.action = na.omit} omitted cases will not appear in
  the residuals, whereas if \code{na.action = na.exclude} they will
  appear (in predictions and standard errors), with residual value
  \code{NA}.  See also \code{\link[stats]{napredict}}.
}
\note{
Variables are first looked for in \code{newdata} and then searched for
  in the usual way (which will include the environment of the formula
  used in the fit).  A warning will be given if the
  variables found are not of the same length as those in \code{newdata}
  if it was supplied.
}
\examples{
require(graphics)

## example from Venables and Ripley (2002, pp. 190-2.)
ldose <- rep(0:5, 2)
numdead <- c(1, 4, 9, 13, 18, 20, 0, 2, 6, 10, 12, 16)
sex <- factor(rep(c("M", "F"), c(6, 6)))
SF <- cbind(numdead, numalive = 20-numdead)
budworm.lg <- glm(SF ~ sex*ldose, family = binomial)
summary(budworm.lg)

plot(c(1,32), c(0,1), type = "n", xlab = "dose",
     ylab = "prob", log = "x")
text(2^ldose, numdead/20, as.character(sex))
ld <- seq(0, 5, 0.1)
lines(2^ld, predict(budworm.lg, data.frame(ldose = ld,
   sex = factor(rep("M", length(ld)), levels = levels(sex))),
   type = "response"))
lines(2^ld, predict(budworm.lg, data.frame(ldose = ld,
   sex = factor(rep("F", length(ld)), levels = levels(sex))),
   type = "response"))
}
\seealso{
\code{\link[stats]{glm}}, \code{\link[stats]{SafePrediction}}
}
\keyword{internal}
