% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm4_summary.R
\name{summary.glm4}
\alias{summary.glm4}
\title{Summarising Generalized Linear Models Using Sparse Matrices}
\usage{
## S3 method for class 'glm4'
summary(object, p.adjust = NULL, dispersion = NULL, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{object}{an object of class "glm4"}

\item{p.adjust}{returns p-values adjusted using one of several methods implemented in `stats::p.adjust`. Defaults to `NULL` for no adjustment, consistent with `stats::glm`.}

\item{dispersion}{the dispersion parameter for the family used. Either a single numerical value or NULL (the default), when it is inferred from object (see `stats::summary.glm()` details).}

\item{correlation}{logical; if `TRUE`, the correlation matrix of the estimated parameters is returned and printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list object of class `c("summary.glm", "summary.glm4")`. See `stats::summary.glm()` for more details on returned components.
}
\description{
Generates a summary of the `glm4` object to evaluate coefficients, standard errors, model fit criteria, etc.
}
\details{
This function is designed to resemble `stats::summary.glm()` as closely as possible. It calculates the (un)scaled variance-covariance matrix from a `MatrixModels::glm4()` object using `Matrix::chol2inv()` and produces a coefficient table for easy inspection of model parameters.
}
